refactor(#17,#21): extract tool router and crawl/search helpers
